+++ types/src/main/kotlin/gg/skytils/hypixel/types/skyblock/member.kt
@@ -19,17 +19,17 @@ class Member(
 class PlayerData(
     val visited_zones: List<String>,
     val last_death: Long,
-    val perks: Map<String, Int> = emptyMap(),
+    val perks: Map<String, Double> = emptyMap(),
     val active_effects: List<PotionEffect>,
     val paused_effects: List<PotionEffect>,
     val temp_stat_buffs: JsonElement, //FIXME
-    val death_count: Int,
+    val death_count: Double,
     val disabled_potion_effects: List<String>,
     val visited_modes: List<String>,
     val unlocked_coll_tiers: List<String>,
     val crafted_generators: List<String>,
     val fastest_target_practice: Double,
-    val fishing_treasure_caught: Int,
+    val fishing_treasure_caught: Double,
     val experience: Map<String, Double>,
 )
 
@@ -37,7 +37,7 @@ class PlayerData(
 class PotionEffect(
     @SerialName("effect")
     val name: String,
-    val level: Int,
+    val level: Double,
     val modifiers: List<PotionModifier> = emptyList(),
     val ticks_remaining: Long = 0,
     val infinite: Boolean
@@ -45,7 +45,7 @@ class PotionEffect(
     @Serializable
     class PotionModifier(
         val name: String,
-        val amp: Int
+        val amp: Double
     )
 }
 
@@ -54,31 +54,31 @@ class AccessoryBagData(
     val tuning: JsonObject,
     val selected_power: String? = null,
     val unlocked_powers: List<String> = emptyList(),
-    val highest_magical_power: Int = 0,
-    val bag_upgrades_purchased: Int = 0
+    val highest_magical_power: Double = 0.0,
+    val bag_upgrades_purchased: Double = 0.0
 )
 
 //TODO: Finish other stuff
 @Serializable
 class PlayerStats(
-    val kills: Map<String, Int>,
-    val deaths: Map<String, Int>,
+    val kills: Map<String, Double>,
+    val deaths: Map<String, Double>,
     val highest_critical_damage: Double,
-    val items_fished: Map<String, Int>,
+    val items_fished: Map<String, Double>,
     val auctionStats: AuctionStats
 )
 
 @Serializable
 class AuctionStats(
-    val bids: Int,
+    val bids: Double,
     val highest_bid: Double,
-    val won: Int,
-    val total_bought: Map<String, Int>,
+    val won: Double,
+    val total_bought: Map<String, Double>,
     val gold_spent: Double,
-    val created: Int,
+    val created: Double,
     val fees: Double,
-    val completed: Int,
-    val total_sold: Map<String, Int>,
+    val completed: Double,
+    val total_sold: Map<String, Double>,
     val gold_earned: Double,
-    val no_bods: Int
+    val no_bods: Double
 )
